initial pass at steps
