+++ .golangci.yml
