+++ examples/examples_test.go
