+++ examples/go.mod
@@ -1,3 +1,3 @@
-module github.com/pulumi/pulumi-xyz/examples
+module github.com/pulumiverse/pulumi-acme/examples
 
 go 1.16
