+++ provider/cmd/pulumi-resource-acme/.gitignore
@@ -1,0 +1,1 @@
+schema.go
