+++ provider/cmd/pulumi-resource-acme/main.go
@@ -12,7 +12,22 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package version
+//go:generate go run ./generate.go
 
-// Version is initialized by the Go linker to contain the semver of this build.
-var Version string
+package main
+
+import (
+	_ "embed"
+
+	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge"
+	acme "github.com/pulumiverse/pulumi-acme/provider"
+	"github.com/pulumiverse/pulumi-acme/provider/pkg/version"
+)
+
+//go:embed schema-embed.json
+var pulumiSchema []byte
+
+func main() {
+	// Modify the path to point to the new provider
+	tfbridge.Main("acme", version.Version, acme.Provider(), pulumiSchema)
+}
