--- provider/cmd/pulumi-resource-xyz/.gitignore
@@ -1,1 +1,0 @@
-schema.go
